test(linux): measure PR260 static symbol address addends
Overlay only the seven production blobs from clean #260 onto the exact 73/100 first-100 checkpoint after #259. Preserve the batch harness and all other product state so movement is attributable solely to canonical signed static relocation addends.

中文说明：仅将 clean #260 的 7 个生产 blob 叠加到 #259 后精确的 73/100 Linux 首批 100 TU checkpoint；保持 batch harness 与其他产品状态不变，使本轮变化只归因于 canonical signed static relocation addend。

+++ src/frontend/ast_verifier.c
@@ -1302,7 +1302,7 @@ bool minic_c0_program_verify_target(const MinicC0Program *program,
                     return false;
                 }
                 {
-                    size_t target_addend;
+                    int64_t target_addend;
 
                     if (!minic_data_layout_global_relocation_target_addend(
                             minic_target_info_data_layout(target),
